requirements.txt
fastapi
uvicorn
pydantic
httpx
openai-whisper @ git+https://github.com/openai/whisper.git
torch
gtts
whisper
python-dotenv
